+++ packages/eslint-plugin/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @digitalsilk/eslint-plugin
 
+## 0.7.0
+
+### Minor Changes
+
+- Cleanup
+
+### Patch Changes
+
+- Updated dependencies
+  - @digitalsilk/babel-preset-default@0.7.0
+  - @digitalsilk/eslint-config@0.7.0
+
 ## 0.6.0
 
 ### Minor Changes